Consolidate item category icons.
